refactor model into mod
